+++ internal/database/ent/db/clan.go
@@ -18,9 +18,9 @@ type Clan struct {
 	// ID of the ent.
 	ID string `json:"id,omitempty"`
 	// CreatedAt holds the value of the "created_at" field.
-	CreatedAt int `json:"created_at,omitempty"`
+	CreatedAt int64 `json:"created_at,omitempty"`
 	// UpdatedAt holds the value of the "updated_at" field.
-	UpdatedAt int `json:"updated_at,omitempty"`
+	UpdatedAt int64 `json:"updated_at,omitempty"`
 	// Tag holds the value of the "tag" field.
 	Tag string `json:"tag,omitempty"`
 	// Name holds the value of the "name" field.
@@ -89,13 +89,13 @@ func (c *Clan) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field created_at", values[i])
 			} else if value.Valid {
-				c.CreatedAt = int(value.Int64)
+				c.CreatedAt = value.Int64
 			}
 		case clan.FieldUpdatedAt:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field updated_at", values[i])
 			} else if value.Valid {
-				c.UpdatedAt = int(value.Int64)
+				c.UpdatedAt = value.Int64
 			}
 		case clan.FieldTag:
 			if value, ok := values[i].(*sql.NullString); !ok {
